Backend/service_users/app/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
databases
alembic
passlib[bcrypt]
PyJWT
pydantic
email-validator
python-multipart
fastapi-pagination

# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-sqlalchemy
